requirements.txt
flask==1.1.2
flask-user==1.0.2.2
python-dotenv==0.15.0
flask-migrate==2.7.0
email-validator==1.1.2
flask-babelex==0.9.4
coinbase-commerce==1.0.1
pyopenssl==20.0.1
SQLAlchemy-JSONField==1.0.0